Remove deprecated edit_configs() function [CLI]
